- adds packages - moves fragment files
