+++ app/src/main/java/com/innoprog/android/feature/auth/presentation/NewPasswordFragment.kt
@@ -1,13 +1,15 @@
-package com.innoprog.android
+package com.innoprog.android.feature.auth.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.innoprog.android.R
+import com.innoprog.android.base.BaseFragment
 import com.innoprog.android.databinding.FragmentNewPasswordBinding
 
-class NewPasswordFragment : BindingFragment<FragmentNewPasswordBinding>() {
+class NewPasswordFragment : BaseFragment<FragmentNewPasswordBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
