+++ app/src/main/res/layout/fragment_authorization.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".AuthorizationFragment">
+    tools:context=".feature.auth.presentation.AuthorizationFragment">
 
     <TextView
         android:id="@+id/tvAuthorizationTitle"
