+++ app/src/main/res/layout/fragment_code_entry.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".CodeEntryFragment">
+    tools:context=".feature.auth.presentation.CodeEntryFragment">
 
     <TextView
         android:id="@+id/tvCodeEntryTitle"
