+++ app/src/main/res/layout/fragment_new_password.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".NewPasswordFragment">
+    tools:context=".feature.auth.presentation.NewPasswordFragment">
 
     <TextView
         android:id="@+id/tvNewPasswordTitle"
