+++ app/src/main/res/layout/fragment_password_recovery.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".PasswordRecoveryFragment">
+    tools:context=".feature.auth.presentation.PasswordRecoveryFragment">
 
     <TextView
         android:id="@+id/tvPasswordRecoveryTitle"
