+++ app/src/main/res/layout/fragment_registration.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".RegistrationFragment">
+    tools:context=".feature.auth.presentation.RegistrationFragment">
 
     <TextView
         android:id="@+id/tvRegistrationTitle"
