+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,8 +7,9 @@
 
     <fragment
         android:id="@+id/authorizationFragment"
-        android:name="com.innoprog.android.AuthorizationFragment"
-        android:label="AuthorizationFragment" >
+        android:name="com.innoprog.android.feature.auth.presentation.AuthorizationFragment"
+        android:label="AuthorizationFragment"
+        tools:layout="@layout/fragment_authorization">
         <action
             android:id="@+id/action_authorizationFragment_to_registrationFragment"
             app:destination="@id/registrationFragment" />
@@ -19,8 +20,9 @@
 
     <fragment
         android:id="@+id/passwordRecoveryFragment"
-        android:name="com.innoprog.android.PasswordRecoveryFragment"
-        android:label="PasswordRecoveryFragment" >
+        android:name="com.innoprog.android.feature.auth.presentation.PasswordRecoveryFragment"
+        android:label="PasswordRecoveryFragment"
+        tools:layout="@layout/fragment_password_recovery">
         <action
             android:id="@+id/action_passwordRecoveryFragment_to_codeEntryFragment"
             app:destination="@id/codeEntryFragment" />
@@ -28,20 +30,23 @@
 
     <fragment
         android:id="@+id/registrationFragment"
-        android:name="com.innoprog.android.RegistrationFragment"
-        android:label="RegistrationFragment" >
+        android:name="com.innoprog.android.feature.auth.presentation.RegistrationFragment"
+        android:label="RegistrationFragment"
+        tools:layout="@layout/fragment_registration">
         <action
             android:id="@+id/action_registrationFragment_to_codeEntryFragment"
             app:destination="@id/codeEntryFragment" />
     </fragment>
+
     <fragment
         android:id="@+id/codeEntryFragment"
-        android:name="com.innoprog.android.CodeEntryFragment"
+        android:name="com.innoprog.android.feature.auth.presentation.CodeEntryFragment"
         android:label="fragment_code_entry"
         tools:layout="@layout/fragment_code_entry" />
+
     <fragment
         android:id="@+id/newPasswordFragment"
-        android:name="com.innoprog.android.NewPasswordFragment"
+        android:name="com.innoprog.android.feature.auth.presentation.NewPasswordFragment"
         android:label="fragment_new_password"
         tools:layout="@layout/fragment_new_password" />
 </navigation>
